fix a butg rudder positon overflow
1. change rudder position calculation into uint32_t
2. add rudder debug function

+++ ESP32/include/DiyActivePedal_types.h
@@ -280,8 +280,8 @@ struct DAP_calculationVariables_st
   float WS_freq;
   bool Rudder_status;
   uint8_t pedal_type;
-  uint16_t sync_pedal_position;
-  uint16_t current_pedal_position;
+  uint32_t sync_pedal_position;
+  uint32_t current_pedal_position;
   float current_pedal_position_ratio;
   float Sync_pedal_position_ratio;
   bool rudder_brake_status;
